requirements.txt
python-telegram-bot[ext]=20.7
